Add 'Test' suffix to the test class names
