+++ selenium/che-selenium-test/src/test/java/org/eclipse/che/selenium/factory/DirectUrlFactoryWithKeepDirectoryTest.java
@@ -10,12 +10,12 @@
  */
 package org.eclipse.che.selenium.factory;
 
+import static org.eclipse.che.selenium.core.constant.TestTimeoutsConstants.UPDATING_PROJECT_TIMEOUT_SEC;
 import static org.testng.Assert.assertTrue;
-import static org.testng.Assert.fail;
 
+import com.google.common.collect.ImmutableList;
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
-import java.util.Arrays;
 import java.util.List;
 import org.eclipse.che.selenium.core.SeleniumWebDriver;
 import org.eclipse.che.selenium.core.client.TestProjectServiceClient;
@@ -27,86 +27,62 @@ import org.eclipse.che.selenium.pageobject.Events;
 import org.eclipse.che.selenium.pageobject.NotificationsPopupPanel;
 import org.eclipse.che.selenium.pageobject.ProjectExplorer;
 import org.eclipse.che.selenium.pageobject.dashboard.Dashboard;
-import org.openqa.selenium.TimeoutException;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-/** @author Musienko Maxim */
-  public class DirectUrlFactoryWithRootFolder {
-  private static final String EXPECTED_PROJECT = "quickstart";
+public class DirectUrlFactoryWithKeepDirectoryTest {
 
   @Inject
   @Named("github.username")
   private String gitHubUsername;
 
   @Inject private ProjectExplorer projectExplorer;
-  @Inject private TestUser testUser;
   @Inject private TestFactoryInitializer testFactoryInitializer;
   @Inject private NotificationsPopupPanel notificationsPopupPanel;
   @Inject private Events events;
   @Inject private SeleniumWebDriver seleniumWebDriver;
   @Inject private TestWorkspaceServiceClient workspaceServiceClient;
   @Inject private TestProjectServiceClient projectServiceClient;
+  @Inject private TestUser testUser;
   @Inject private Dashboard dashboard;
 
-  private TestFactory testFactoryWithRootFolder;
+  private TestFactory testFactoryWithKeepDir;
 
   @BeforeClass
   public void setUp() throws Exception {
-    testFactoryWithRootFolder =
-        testFactoryInitializer.fromUrl("https://github.com/" + gitHubUsername + "/quickstart");
+    testFactoryWithKeepDir =
+        testFactoryInitializer.fromUrl(
+            "https://github.com/" + gitHubUsername + "/gitPullTest/tree/master/my-lib");
   }
 
   @AfterClass
   public void tearDown() throws Exception {
-    testFactoryWithRootFolder.delete();
+    testFactoryWithKeepDir.delete();
   }
 
   @Test
-  public void factoryWithDirectUrlWithRootFolder() throws Exception {
-    String expectedMessInTheEventsPanel = "Project " + EXPECTED_PROJECT + " imported";
-    List<String> expectedItemsAfterCloning =
-        Arrays.asList(
-            "CHANGELOG.md",
-            "Dockerfile",
-            "LICENSE",
-            "README.md",
-            "favicon.ico",
-            "index.html",
-            "karma-test-shim.js",
-            "karma.conf.js",
-            "package.json",
-            "protractor.config.js",
-            "styles.css",
-            "systemjs.config.js",
-            "tsconfig.json",
-            "tslint.json",
-            "typings.json",
-            "wallaby.js");
-
-    testFactoryWithRootFolder.authenticateAndOpen();
-
+  public void factoryWithDirectUrlWithKeepDirectory() throws Exception {
+    testFactoryWithKeepDir.authenticateAndOpen();
     projectExplorer.waitProjectExplorer();
-    projectExplorer.waitItem(EXPECTED_PROJECT);
     notificationsPopupPanel.waitProgressPopupPanelClose();
     events.clickEventLogBtn();
+    events.waitExpectedMessage("Project gitPullTest imported", UPDATING_PROJECT_TIMEOUT_SEC);
 
-    try {
-      events.waitExpectedMessage(expectedMessInTheEventsPanel);
-    } catch (TimeoutException ex) {
-      // remove try-catch block after issue has been resolved
-      fail("Known issue https://github.com/eclipse/che/issues/6440");
-    }
-    projectExplorer.openItemByPath(EXPECTED_PROJECT);
+    projectExplorer.waitItem("gitPullTest");
+    projectExplorer.quickExpandWithJavaScript();
+    projectExplorer.waitItem("gitPullTest/my-lib/pom.xml");
 
-    String currentWsId =
+    String wsId =
         workspaceServiceClient
             .getByName(seleniumWebDriver.getWorkspaceNameFromBrowserUrl(), testUser.getName())
             .getId();
+
     List<String> visibleItems = projectExplorer.getNamesOfAllOpenItems();
-    assertTrue(visibleItems.containsAll(expectedItemsAfterCloning));
-    String currentTypeOfProject = projectServiceClient.getFirstProject(currentWsId).getType();
-    assertTrue(currentTypeOfProject.equals("blank"));
+    assertTrue(
+        visibleItems.containsAll(ImmutableList.of("gitPullTest", "my-lib", "src", "pom.xml")));
+
+    String projectType = projectServiceClient.getFirstProject(wsId).getType();
+    assertTrue(projectType.equals("blank"));
   }
 }
