+++ selenium/che-selenium-test/src/test/java/org/eclipse/che/selenium/factory/DirectUrlFactoryWithRootFolderTest.java
@@ -10,13 +10,12 @@
  */
 package org.eclipse.che.selenium.factory;
 
-import static org.eclipse.che.selenium.core.constant.TestTimeoutsConstants.UPDATING_PROJECT_TIMEOUT_SEC;
 import static org.testng.Assert.assertTrue;
 import static org.testng.Assert.fail;
 
-import com.google.common.collect.ImmutableList;
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
+import java.util.Arrays;
 import java.util.List;
 import org.eclipse.che.selenium.core.SeleniumWebDriver;
 import org.eclipse.che.selenium.core.client.TestProjectServiceClient;
@@ -25,23 +24,22 @@ import org.eclipse.che.selenium.core.factory.TestFactory;
 import org.eclipse.che.selenium.core.factory.TestFactoryInitializer;
 import org.eclipse.che.selenium.core.user.TestUser;
 import org.eclipse.che.selenium.pageobject.Events;
-import org.eclipse.che.selenium.pageobject.Ide;
 import org.eclipse.che.selenium.pageobject.NotificationsPopupPanel;
 import org.eclipse.che.selenium.pageobject.ProjectExplorer;
-import org.eclipse.che.selenium.pageobject.PullRequestPanel;
-import org.openqa.selenium.NoSuchElementException;
+import org.eclipse.che.selenium.pageobject.dashboard.Dashboard;
 import org.openqa.selenium.TimeoutException;
-import org.openqa.selenium.WebDriverException;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-public class DirectUrlFactoryWithSpecificBranch {
+/** @author Musienko Maxim */
+public class DirectUrlFactoryWithRootFolderTest {
+  private static final String EXPECTED_PROJECT = "quickstart";
+
   @Inject
   @Named("github.username")
   private String gitHubUsername;
 
-  @Inject private Ide ide;
   @Inject private ProjectExplorer projectExplorer;
   @Inject private TestUser testUser;
   @Inject private TestFactoryInitializer testFactoryInitializer;
@@ -49,63 +47,66 @@ public class DirectUrlFactoryWithSpecificBranch {
   @Inject private Events events;
   @Inject private SeleniumWebDriver seleniumWebDriver;
   @Inject private TestWorkspaceServiceClient workspaceServiceClient;
-  @Inject private TestProjectServiceClient testProjectServiceClient;
-  @Inject private PullRequestPanel pullRequestPanel;
+  @Inject private TestProjectServiceClient projectServiceClient;
+  @Inject private Dashboard dashboard;
 
-  private TestFactory testFactoryWithSpecificBranch;
+  private TestFactory testFactoryWithRootFolder;
 
   @BeforeClass
   public void setUp() throws Exception {
-    testFactoryWithSpecificBranch =
-        testFactoryInitializer.fromUrl(
-            "https://github.com/" + gitHubUsername + "/gitPullTest/tree/contrib-12092015");
+    testFactoryWithRootFolder =
+        testFactoryInitializer.fromUrl("https://github.com/" + gitHubUsername + "/quickstart");
   }
 
   @AfterClass
   public void tearDown() throws Exception {
-    if (workspaceServiceClient.exists(gitHubUsername, testUser.getName())) {
-      testFactoryWithSpecificBranch.delete();
-    }
+    testFactoryWithRootFolder.delete();
   }
 
   @Test
-  public void factoryWithDirectUrlWithSpecificBranch() throws Exception {
-    try {
-      testFactoryWithSpecificBranch.authenticateAndOpen();
-    } catch (NoSuchElementException ex) {
-      // remove try-catch block after issue has been resolved
-      fail("https://github.com/eclipse/che/issues/8671");
-    }
+  public void factoryWithDirectUrlWithRootFolder() throws Exception {
+    String expectedMessInTheEventsPanel = "Project " + EXPECTED_PROJECT + " imported";
+    List<String> expectedItemsAfterCloning =
+        Arrays.asList(
+            "CHANGELOG.md",
+            "Dockerfile",
+            "LICENSE",
+            "README.md",
+            "favicon.ico",
+            "index.html",
+            "karma-test-shim.js",
+            "karma.conf.js",
+            "package.json",
+            "protractor.config.js",
+            "styles.css",
+            "systemjs.config.js",
+            "tsconfig.json",
+            "tslint.json",
+            "typings.json",
+            "wallaby.js");
+
+    testFactoryWithRootFolder.authenticateAndOpen();
 
     projectExplorer.waitProjectExplorer();
+    projectExplorer.waitItem(EXPECTED_PROJECT);
     notificationsPopupPanel.waitProgressPopupPanelClose();
     events.clickEventLogBtn();
 
-    events.waitExpectedMessage("Project gitPullTest imported", UPDATING_PROJECT_TIMEOUT_SEC);
-    events.waitExpectedMessage(
-        "Successfully configured and cloned source code of gitPullTest.",
-        UPDATING_PROJECT_TIMEOUT_SEC);
-    events.waitExpectedMessage(
-        "Project: gitPullTest | cloned from: gitPullTest | remote branch: refs/remotes/origin/contrib-12092015 | local branch: contrib-12092015",
-        UPDATING_PROJECT_TIMEOUT_SEC);
+    try {
+      events.waitExpectedMessage(expectedMessInTheEventsPanel);
+    } catch (TimeoutException ex) {
+      // remove try-catch block after issue has been resolved
+      fail("Known issue https://github.com/eclipse/che/issues/6440");
+    }
+    projectExplorer.openItemByPath(EXPECTED_PROJECT);
 
-    projectExplorer.selectItem("gitPullTest");
-    pullRequestPanel.waitOpenPanel();
-
-    projectExplorer.expandPathInProjectExplorer("gitPullTest/my-lib");
-    projectExplorer.waitItem("gitPullTest/my-lib/pom.xml");
-
-    String wsId =
+    String currentWsId =
         workspaceServiceClient
             .getByName(seleniumWebDriver.getWorkspaceNameFromBrowserUrl(), testUser.getName())
             .getId();
-
     List<String> visibleItems = projectExplorer.getNamesOfAllOpenItems();
-    assertTrue(
-        visibleItems.containsAll(
-            ImmutableList.of("gitPullTest", "my-lib", "my-webapp", "src", "pom.xml")));
-
-    String projectType = testProjectServiceClient.getFirstProject(wsId).getType();
-    assertTrue(projectType.equals("blank"));
+    assertTrue(visibleItems.containsAll(expectedItemsAfterCloning));
+    String currentTypeOfProject = projectServiceClient.getFirstProject(currentWsId).getType();
+    assertTrue(currentTypeOfProject.equals("blank"));
   }
 }
