+++ selenium/che-selenium-test/src/test/resources/suites/CheOneThreadTestsSuite.xml
@@ -75,9 +75,9 @@
             <class name="org.eclipse.che.selenium.git.RevertCommitTest"/>
             <class name="org.eclipse.che.selenium.git.WorkingWithPullConflictsTest"/>
             <class name="org.eclipse.che.selenium.dashboard.ImportProjectFromGitHubTest"/>
-            <class name="org.eclipse.che.selenium.factory.DirectUrlFactoryWithKeepDirectory"/>
-            <class name="org.eclipse.che.selenium.factory.DirectUrlFactoryWithRootFolder"/>
-            <class name="org.eclipse.che.selenium.factory.DirectUrlFactoryWithSpecificBranch"/>
+            <class name="org.eclipse.che.selenium.factory.DirectUrlFactoryWithKeepDirectoryTest"/>
+            <class name="org.eclipse.che.selenium.factory.DirectUrlFactoryWithRootFolderTest"/>
+            <class name="org.eclipse.che.selenium.factory.DirectUrlFactoryWithSpecificBranchTest"/>
         </classes>
     </test>
 </suite>
